+++ common/rmi/src/main/java/is/codion/common/rmi/server/DefaultServerAdmin.java
@@ -84,7 +84,7 @@ public class DefaultServerAdmin extends UnicastRemoteObject implements ServerAdm
 
 	@Override
 	public final String systemProperties() {
-		return PropertyStore.systemProperties();
+		return configuration.systemProperties() ? PropertyStore.systemProperties() : "";
 	}
 
 	@Override